Separate timer/throttle tests and adapt timer tests for Jest fake timers: https://jestjs.io/blog/2020/05/05/jest-26#new-fake-timers
