+++ src/util.test.js
@@ -1,89 +1,6 @@
 const jQuery = require('jquery');
 global.jQuery = jQuery;
 
-describe('throttle function executes given function ', () => {
-
-    test('once immediatelly and once after given wait interval when {leading: false} is not provided', async () => {
-        expect.assertions(3);
-
-        const wait = 500;
-        const mockFunction = jest.fn();
-
-        const throttledFunction = Oskari.util.throttle(mockFunction, wait);
-
-        throttledFunction();
-        expect(mockFunction.mock.calls.length).toEqual(1);
-        throttledFunction();
-        expect(mockFunction.mock.calls.length).toEqual(1);
-        await sleep(wait);
-        expect(mockFunction.mock.calls.length).toEqual(2);
-    });
-
-    test('zero times immediatelly and once after given wait interval when {leading: false} is provided', async () => {
-        expect.assertions(2);
-
-        jest.setTimeout(7000);
-
-        const functionCallCount = 5;
-        const wait = 500;
-        const mockFunction = jest.fn();
-
-        const throttledFunction = Oskari.util.throttle(mockFunction, wait, { leading: false });
-
-        for (let i = 0; i < functionCallCount; i++) {
-            throttledFunction();
-            if (i === 0) {
-                expect(mockFunction.mock.calls.length).toEqual(0);
-            }
-        };
-        /* Sleep enough to verify that only one function call within given wait time is set to be executed in future
-        *  (multiple throttle function calls received as burst within given wait- parameter do not spread all given function calls to be executed in future with interval of wait- parameter)
-        */
-        await sleep(wait * functionCallCount);
-        expect(mockFunction.mock.calls.length).toEqual(1);
-    });
-
-    test('once immediatelly and zero time after given wait interval when {trailing: false} is provided', async () => {
-        expect.assertions(2);
-
-        const functionCallCount = 2;
-        const wait = 500;
-        const mockFunction = jest.fn();
-
-        const throttledFunction = Oskari.util.throttle(mockFunction, wait, { trailing: false });
-
-        for (let i = 0; i < functionCallCount; i++) {
-            throttledFunction();
-            if (i === 0) {
-                expect(mockFunction.mock.calls.length).toEqual(1);
-            }
-        };
-        await sleep(wait);
-        expect(mockFunction.mock.calls.length).toEqual(1);
-    });
-
-    test('immediatelly without throttling when called with interval >= throttle wait parameter', async () => {
-
-        const functionCallCount = 5;
-        expect.assertions(functionCallCount);
-
-        const wait = 200;
-        const mockFunction = jest.fn();
-
-        const throttledFunction = Oskari.util.throttle(mockFunction, wait);
-
-        for (let i = 0; i < functionCallCount; i++) {
-            throttledFunction();
-            await sleep(wait);
-            expect(mockFunction.mock.calls.length).toEqual(i + 1);
-        };
-    });
-
-    function sleep(ms) {
-        return new Promise(resolve => setTimeout(resolve, ms));
-    }
-});
-
 describe('isNumber function', () => {
 
     test('returns true when parameter is number', () => {
